+++ src/pdesc.h
@@ -181,7 +181,7 @@ void            remove_proxy_desc_rings(proxy_desc_t *cur);
 forward_desc_t* create_fwd_desc(uint16_t seq_no, uint32_t data_len, char *data);
 
 int             queue_packet(int sock_fd, proxy_desc_t *cur, char *buf, size_t bufsiz,
-                             uint32_t dest_ip, u_int16_t dest_port, uint32_t state);
+                             uint32_t dest_ip, uint32_t dest_port, uint32_t state);
 
 uint32_t        send_packets(forward_desc_t *ring[], int *xfer_idx, int *await_send, int *sock, uint16_t window_size);
